+++ Acessibilidade_MC3/Acessibilidade_MC3/Controller/EmpresasViewController.swift
@@ -69,94 +69,6 @@ class EmpresasViewController: UIViewController {
     }
 }
 
-class Empresa {
-    var nome: String
-    var localizacao: String
-    var site: String?
-    var nota: Float = 0.0
-    var recomendacao: Int = 0
-    var acessibilidade: [Acessibilidade] = []
-    var avaliacoes: [Avaliacao] = []
-    
-    init(nome: String, localizacao: String, site: String?) {
-        self.nome = nome
-        self.localizacao = localizacao
-        
-        if let site = site {
-            self.site = site
-        }
-    }
-    
-    public func adicionaAvaliacao(avaliacao: Avaliacao) {
-        self.avaliacoes.append(avaliacao)
-        self.calculaMediaNota()
-        self.calculaPorcentagemRecomendacao()
-        self.registraAcessibilidade(avaliacao: avaliacao)
-    }
-    
-    private func calculaMediaNota() {
-        var media: Float = 0
-        for avaliacao in avaliacoes {
-            media += avaliacao.nota
-        }
-        
-        if avaliacoes.count > 0 {
-            media /= Float(avaliacoes.count)
-        }
-        self.nota = media
-    }
-    
-    private func calculaPorcentagemRecomendacao() {
-        var recomendacoes: Int = 0
-        for avaliacao in avaliacoes {
-            if avaliacao.recomendacao {
-                recomendacoes += 1
-            }
-        }
-        
-        if avaliacoes.count > 0 {
-            recomendacoes = recomendacoes * 100 / avaliacoes.count
-        }
-        
-        self.recomendacao = recomendacoes
-    }
-    
-    private func registraAcessibilidade(avaliacao: Avaliacao) {
-        for acesso in avaliacao.acessibilidade {
-            if !self.acessibilidade.contains(acesso) {
-                self.acessibilidade.append(acesso)
-            }
-        }
-    }
-    
-}
-
-class Avaliacao {
-    init(data: Date, titulo: String, vantagens: String, desvantagens: String, sugestoes: String?, cargo: Cargo, nota: Float, recomendacao: Bool, acessibilidade: [Acessibilidade]) {
-        self.data = data
-        self.titulo = titulo
-        self.vantagens = vantagens
-        self.desvantagens = desvantagens
-        self.sugestoes = sugestoes
-        self.cargo = cargo
-        self.nota = nota
-        self.recomendacao = recomendacao
-        self.acessibilidade = acessibilidade
-        
-    }
-    
-    var data: Date
-    var titulo: String
-    var vantagens: String
-    var desvantagens: String
-    var sugestoes: String?
-    var cargo: Cargo
-    var nota: Float
-    var recomendacao: Bool
-    var acessibilidade: [Acessibilidade]
-    
-}
-
 enum Acessibilidade: String {
     case deficienciaMotora = "SIA"
     case deficienciaVisual = "SIDV"
