+++ Acessibilidade_MC3/Acessibilidade_MC3/View/AvaliacaoHeaderView.swift
@@ -1,0 +1,22 @@
+//
+//  AvaliacaoHeaderView.swift
+//  Acessibilidade_MC3
+//
+//  Created by Lia Kassardjian on 02/09/19.
+//  Copyright © 2019 Lia Kassardjian. All rights reserved.
+//
+
+import UIKit
+
+class AvaliacaoHeaderView: UITableViewHeaderFooterView {
+
+    @IBOutlet weak var tituloLabel: UILabel!
+    /*
+    // Only override draw() if you perform custom drawing.
+    // An empty implementation adversely affects performance during animation.
+    override func draw(_ rect: CGRect) {
+        // Drawing code
+    }
+    */
+
+}
